cleanup

+++ clibs/rmlui/render.cpp
@@ -51,20 +51,20 @@ static uint32_t getTextureFlags(Rml::SamplerFlag flags) {
     }
 }
 
-void Renderer::RenderGeometry(Rml::Vertex* vertices, int num_vertices, Rml::Index* indices, int num_indices, Rml::TextureHandle texture, Rml::SamplerFlag flags) {
+void Renderer::RenderGeometry(Rml::Vertex* vertices, size_t num_vertices, Rml::Index* indices, size_t num_indices, Rml::TextureHandle texture, Rml::SamplerFlag flags) {
     BGFX(encoder_set_state)(mEncoder, RENDER_STATE, 0);
     bgfx_transient_vertex_buffer_t tvb;
-    BGFX(alloc_transient_vertex_buffer)(&tvb, num_vertices, (bgfx_vertex_layout_t*)mcontext->layout);
+    BGFX(alloc_transient_vertex_buffer)(&tvb, (uint32_t)num_vertices, (bgfx_vertex_layout_t*)mcontext->layout);
 
     memcpy(tvb.data, vertices, num_vertices * sizeof(Rml::Vertex));
-    BGFX(encoder_set_transient_vertex_buffer)(mEncoder, 0, &tvb, 0, num_vertices);
+    BGFX(encoder_set_transient_vertex_buffer)(mEncoder, 0, &tvb, 0, (uint32_t)num_vertices);
 
     bgfx_transient_index_buffer_t tib;
-    BGFX(alloc_transient_index_buffer)(&tib, num_indices, true);
+    BGFX(alloc_transient_index_buffer)(&tib, (uint32_t)num_indices, true);
 
     static_assert(sizeof(Rml::Index) == sizeof(uint32_t));
     memcpy(tib.data, indices, num_indices * sizeof(Rml::Index));
-    BGFX(encoder_set_transient_index_buffer)(mEncoder, &tib, 0, num_indices);
+    BGFX(encoder_set_transient_index_buffer)(mEncoder, &tib, 0, (uint32_t)num_indices);
 
     auto fe = FE(texture);
     auto get_shader = [&](){
